debug cout enabled

+++ src/graphlab/graph/ingress/ingress_edge_decision.hpp
@@ -39,8 +39,8 @@ namespace graphlab {
     std::cout << "coords_pair_dist:" << topology_to_str(src) << " to " << topology_to_str(dst) << std::endl;
     size_t dist = 0;
     for (size_t j = 0; j < 3; ++j) {
-    	size_t src_val = (src >> (10 - 5 * j)) & 0x1F;
-    	size_t dst_val = (dst >> (10 - 5 * j)) & 0x1F;
+    	int src_val = (int)(src >> (10 - 5 * j)) & 0x1F;
+    	int dst_val = (int)(dst >> (10 - 5 * j)) & 0x1F;
     	size_t absval = std::abs(src_val - dst_val);
     	dist += std::min(absval, 23 - absval);
       std::cout << "j=" << j << ", src_val=" << src_val << ", dst_val=" << dst_val << ", absval=" << absval << ", dist=" << dist << std::endl;
